sbomnix: delete legacy fallback store
Remove the fallback component store and the batched find_derivers helper that
existed only for nix-store -qd based deriver lookup. find_deriver now relies
on structured nix derivation show data and reports unloadable derivers as a
hard failure.

Drop fallback-only batching tests and add argv coverage proving find_deriver
does not issue any unexpected legacy command for unloadable structured
deriver metadata.
